Refactor the client store endpoint so it returns a standarized response.
